flask-mongo/requirements.txt
Flask==2.0.2
pymongo==4.4.1
Flask-Cors==3.0.10
Werkzeug==2.2.2
PyYAML
coveragepy